requirements.txt
# Grocery Genie Dependencies
psycopg2-binary  # PostgreSQL database connection
requests         # HTTP requests for APIs
PyYAML          # YAML file processing
python-dotenv   # Environment variable loading
beautifulsoup4  # HTML parsing for Walmart data
pytest          # Testing framework for CI/CD
ruff            # Python linter and formatter